unified setup and start server with saved steps
The new command (python) hdd-indexer.py starts the setup if it has not been completed. It automatically starts the default web browser at localhost:8000 to complete the rest of the setup, or if it has been completed, start the homepage. The setup now stores its states in a file called .setup.py which is a pickled dictionary.

Closes #7 save setup status and steps

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>