fixes #17 opensub CannotSendRequest error
Used a function.dict stored lock to allow only one thread to connect to OpenSub at a time. Also moved the token code to the function init code where sub is instantiated. This allows the entire load operation to be completed with just one token instead of individual token for every request.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>